+++ lib/src/commands/create_command.dart
@@ -1,5 +1,6 @@
 import 'package:args/command_runner.dart';
 import 'package:mason_logger/mason_logger.dart';
+import 'package:moarch/src/commands/create/create_empty_factories_command.dart';
 import 'package:moarch/src/commands/create/create_feature_command.dart';
 import 'package:moarch/src/commands/create/create_model_command.dart';
 
@@ -9,6 +10,7 @@ class CreateCommand extends Command<int> {
   CreateCommand({required Logger logger}) : _logger = logger {
     addSubcommand(CreateFeatureCommand(logger: logger));
     addSubcommand(CreateModelCommand(logger: logger));
+    addSubcommand(CreateEmptyFactoriesCommand(logger: logger));
   }
 
   final Logger _logger;